Move store-location ruby stuff into their own folders/files
