+++ app/null_objects/null_geocoded_result.rb
@@ -1,0 +1,5 @@
+class NullGeocodedResult < NullObject
+  def coordinates
+    [0.0, 0.0]
+  end
+end
